# back/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
...
        "/kermesse/{kermesse_id}/stands": {
            "get": {
                "description": "Récupère tous les stands associés à une kermesse spécifique",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stands"
                ],
                "summary": "Récupérer tous les stands pour une kermesse",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Kermesse ID",
                        "name": "kermesse_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des stands",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Stand"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid kermesse ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Could not retrieve stands",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/parents": {
            "post": {
                "description": "Créer une nouvelle entrée parent",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "parents"
                ],
                "summary": "Créer un parent",
                "parameters": [
                    {
                        "description": "Parent data",
                        "name": "parent",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Parent"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Parent created",
                        "schema": {
                            "$ref": "#/definitions/models.Parent"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/parents/{id}": {
            "get": {
                "description": "Récupérer un parent par son ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "parents"
                ],
                "summary": "Trouver un parent par ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Parent ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Parent found",
                        "schema": {
                            "$ref": "#/definitions/models.Parent"
                        }
                    },
                    "400": {
                        "description": "Invalid ID format",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Parent not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Mettre à jour les informations d'un parent",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "parents"
                ],
                "summary": "Mettre à jour un parent",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Parent ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Parent data",
                        "name": "parent",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Parent"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Parent updated",
                        "schema": {
                            "$ref": "#/definitions/models.Parent"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Parent not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Supprimer un parent par son ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "parents"
                ],
                "summary": "Supprimer un parent",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Parent ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Parent deleted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid ID format",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Parent not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/stands": {
            "post": {
                "description": "Crée un stand pour une kermesse",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stands"
                ],
                "summary": "Créer un nouveau stand",
                "parameters": [
                    {
                        "description": "Stand à créer",
                        "name": "stand",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Stand"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Stand créé",
                        "schema": {
                            "$ref": "#/definitions/models.Stand"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Could not create stand",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/stands/{id}": {
            "get": {
                "description": "Récupère un stand spécifique par son ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stands"
                ],
                "summary": "Récupérer un stand par ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Stand ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Stand trouvé",
                        "schema": {
                            "$ref": "#/definitions/models.Stand"
                        }
                    },
                    "400": {
                        "description": "Invalid stand ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Stand not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Supprime un stand spécifique par son ID",
                "tags": [
                    "stands"
                ],
                "summary": "Supprimer un stand par ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Stand ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Stand supprimé",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid stand ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Stand not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Could not delete stand",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Parent": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "tokens_amount_available": {
                    "description": "Solde de jetons disponibles",
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user_id": {
                    "description": "Référence vers la table User",
                    "type": "integer"
                }
            }
        },
...
        "models.Stand": {
            "type": "object",
            "required": [
                "kermesse_id",
                "name",
                "participation_cost",
                "stand_type",
                "stock"
            ],
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "kermesse_id": {
                    "description": "Référence vers la table Kermesse",
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "participation_cost": {
                    "description": "Coût en jetons pour participer",
                    "type": "integer"
                },
                "stand_type": {
                    "description": "ENUM: 'nourriture', 'boisson', 'activité'",
                    "type": "string"
                },
                "stock": {
                    "description": "Stock pour les stands de nourriture ou boisson",
                    "type": "integer"
                },
                "teneur_id": {
                    "description": "Référence vers la table User (teneur de stand)",
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Bearer token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}